Implement second layer of packaging
So now every class is in a package
