+++ src/main/java/duke/command/ExitCommand.java
@@ -1,10 +1,10 @@
-package command;
+package duke.command;
 
-import task.TaskList;
-import textual.Ui;
+import duke.task.TaskList;
+import duke.ui.Ui;
 
 /**
- * Class representing a command to exit Duke.
+ * Class representing a duke.command to exit Duke.
  */
 public class ExitCommand extends Command {
     /**
@@ -16,10 +16,10 @@ public class ExitCommand extends Command {
     }
 
     /**
-     * Executes this command on the given task list and user interface.
+     * Executes this duke.command on the given duke.task list and user interface.
      *
-     * @param tl The task list.
-     * @param ui The user interface displaying events on the task list.
+     * @param tl The duke.task list.
+     * @param ui The user interface displaying events on the duke.task list.
      */
     public void execute(TaskList tl, Ui ui) {
         ui.printMessage("Bye. Hope to see you again soon!");
