+++ src/main/java/duke/task/Event.java
@@ -1,4 +1,4 @@
-package task;
+package duke.task;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
