+++ src/main/java/duke/task/Todo.java
@@ -1,7 +1,7 @@
-package task;
+package duke.task;
 
 /**
- * Class representing a task to be done, but not at any particular time.
+ * Class representing a duke.task to be done, but not at any particular time.
  */
 public class Todo extends Task {
     /**
