+++ src/main/java/duke/ui/Ui.java
@@ -1,4 +1,4 @@
-package textual;
+package duke.ui;
 
 /**
  * Class that handles messages shown to the user of this application.
